+++ src/lib.rs
@@ -33,10 +33,46 @@
 //! assert_eq!(arr, [1.0, 4.0, 6.0, -3.0, 6.0]);
 //! ```
 //!
+//! # Zip:
+//!
+//! ```rust
+//! use higher_order_functions::Zip;
+//!
+//! let a = [1, 2, 3];
+//! let b = ["a", "b", "c"];
+//!
+//! let arr = a.zip(b, |ax, bx| (ax, bx));
+//!
+//! assert_eq!(arr, [(1, "a"), (2, "b"), (3, "c")]);
+//! ```
+//!
+//! ```rust
+//! use higher_order_functions::Zip;
+//!
+//! let a = [1, 2, 3];
+//! let b = [4, 5, 6];
+//!
+//! let arr = a.zip(b, |ax, bx| ax * bx);
+//!
+//! assert_eq!(arr, [4, 10, 18]);
+//! ```
+//!
+//! # Section:
+//!
+//! ```rust
+//! use higher_order_functions::Section;
+//!
+//! let a: [u32; 8] = [1, 2, 3, 4, 5, 6, 7, 8];
+//!
+//! let arr: [u32; 4] = a.section(3); // Extracts 4 elements starting at a[3]
+//!
+//! assert_eq!(arr, [4, 5, 6, 7]);
+//! ```
+//!
 //! To use this, add it as a dependency to your Cargo.toml:
 //! ```toml
 //! [dependencies]
-//! higher_order_functions = "0.1.0"
+//! higher_order_functions = "^0.1.1"
 //! ```
 
 #![no_std]
@@ -47,15 +83,19 @@
 #![feature(maybe_uninit_uninit_array)]
 #![feature(maybe_uninit_extra)]
 
+#![doc(html_root_url = "https://docs.rs/higher_order_functions/0.1.1")]
+
 #[cfg(feature = "std")]
 extern crate std as lib;
 
 #[cfg(all(not(feature = "std"), feature = "alloc"))]
 extern crate alloc as lib;
 
-mod type_equals; // FIXME (#20041): Replace this workaround with real type equality constraints
+mod helpers;
 mod init;
 mod map;
+mod zip;
+mod section;
 
 pub use {
 	init::{
@@ -63,6 +103,12 @@ pub use {
 	},
 	map::{
 		Map,
+	},
+	zip::{
+		Zip,
+	},
+	section::{
+		Section,
 	},
 };