+++ src/type_equals.rs
@@ -1,3 +1,0 @@
-pub trait TypeEquals<T>: From<T> { }
-
-impl<T> TypeEquals<T> for T { }