+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 This project follows [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.1.1](https://crates.io/crates/higher_order_functions/0.1.1) - 2020-09-08
+
+### Added:
+* `Section` trait for extracting fixed-size sections of an array-like structure
+* `Zip` trait for zipping array-like structure using a combining function
+
 ## [v0.1.0](https://crates.io/crates/higher_order_functions/0.1.0) - 2020-09-03
 
 ### Added: